requirements.txt
Django==1.5.4
argparse==1.2.1
distribute==0.6.34
django-tastypie==0.10.0
mimeparse==0.1.3
python-dateutil==2.1
six==1.4.1
wsgiref==0.1.2
